feat: create space metadata correctly
