+++ libs/jwst/src/workspaces/mod.rs
@@ -3,7 +3,7 @@ mod plugins;
 mod transaction;
 mod workspace;
 
-use super::{constants, error, info, trace, warn, JwstResult, Space};
+use super::{constants, error, info, trace, warn, JwstError, JwstResult, Space};
 use plugins::PluginMap;
 
 pub use metadata::WorkspaceMetadata;
